Filters urgent fixes - correct Filter Default QueryString (#2188)
* Fixes for QueryString & Centralising Methods

* Tweaks for Filters
